fix(anomaly): cast err_rate literals to float8 to match sqlx decoding
avg(CASE ... THEN 1.0 ELSE 0.0 END) 在 PG 里返回 NUMERIC 类型，
而 Rust 端 CenterRow::err_rate_avg / GroupRow::err_rate 声明为
Option<f64>，sqlx 解码时拒绝 NUMERIC → FLOAT8 隐式转换。其他
avg 字段（duration / ttft / queue / input）因原始列是 int4/int8
也产生 NUMERIC，但 sqlx 在该路径上接受；CASE 表达式的 1.0/0.0
字面量是 numeric，触发更严的匹配。

CASE 字面量显式 ::float8 后 avg 输出 FLOAT8，解码通过。

+++ boom-gateway/boom-dashboard/src/handlers_admin.rs
@@ -5071,7 +5071,7 @@ pub async fn get_anomalies(
                                  THEN COALESCE(cached_tokens, 0)::float / input_tokens
                                  ELSE NULL END
                  ) AS hit_rate_p75,
-                 avg(CASE WHEN status_code != 200 THEN 1.0 ELSE 0.0 END) AS err_rate_avg
+                 avg(CASE WHEN status_code != 200 THEN 1.0::float8 ELSE 0.0::float8 END) AS err_rate_avg
                FROM boom_request_log
                WHERE created_at >= $1 AND created_at < $2"#,
         )
@@ -5094,7 +5094,7 @@ pub async fn get_anomalies(
                  avg(CASE WHEN input_tokens > 0
                           THEN COALESCE(cached_tokens, 0)::float / input_tokens
                           ELSE NULL END) AS hit_rate_avg,
-                 avg(CASE WHEN status_code != 200 THEN 1.0 ELSE 0.0 END) AS err_rate
+                 avg(CASE WHEN status_code != 200 THEN 1.0::float8 ELSE 0.0::float8 END) AS err_rate
                FROM boom_request_log
                WHERE created_at >= $1 AND created_at < $2
                  AND {dim_col} IS NOT NULL
